reduced the bytesizes of the uuid from 26 to 16

+++ http-server-leveldb-filestorage/src/api/group/delete.rs
@@ -12,7 +12,7 @@ use std::process::exit;
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct Info {
-    priority: u32,
+    priority: u16,
 }
 
 const ROUTE: &'static str = "DEL /api/group";
